requirements.txt
python-telegram-bot==13.7
requests==2.25.1
Flask==2.1.0
gunicorn==20.1.0